Fix scheduler tab summary selectors for Backdrop CMS
The tab summary for scheduled publishing/unpublishing was always
displaying "Not scheduled" due to incorrect CSS selectors targeting
non-existent elements from the Drupal 7 version.

- Replace #edit-publish-on with #edit-publish-on-date
- Replace #edit-unpublish-on with #edit-unpublish-on-date
- Remove references to non-existent datepicker popup selectors
  (#edit-publish-on-datepicker-popup-0 and #edit-unpublish-on-datepicker-popup-0)

The tab summary now correctly displays "Scheduled for publishing",
"Scheduled for unpublishing", or "Not scheduled" based on the actual
state of the date inputs.

Fixes Issue #28